Reworked the job/dag types to remove the job files from the jobs themselves and only carry one copy of each file per dag. Keeping a copy of the JSDL for every job adds tremendous overhead due to (de)serialization of the files. When files are used repeatedly in a dag, we only need one copy.
The WorkflowJobInformationType still has a field for the JSDL, but it's no longer required. We use this for transmission to the DAGRunner when we retrieve new jobs from the database.

Also changed a few SQL statements to try to reduce database overhead from implicit joining. Particularly in getNewJobs(), the statement to find jobs with no outstanding dependencies was incredibly slow (guessing at least polynomial in the size of the dag). Now, jobs in the database have an extra field "parentDeps" which is the number of parents that still have to run. Getting a new job just means looking for any jobs that have a "0" in that field and a status of "ready" or "waiting". To make this work, we added code to setJobStatus() so whenever a job goes into "finished" state, the "parentDeps" fields for all its children are decremented.