Support notifications (both persistent and non-persistent) with attachments.
PublisherTopic.publish() takes an attachment argument, which it passes
to WSNotificationContainerService.

WSNotificationContainerService takes an attachment argument.
For non-persistent notifications, it passes the attachment to the Actor.
For persistent notifications, it passes the attachment to
PersistentOutcallContainerService.schedule().

NotificationOutcallActor.enactOutcall() takes an attachment argument,
which it sends with the outcall.

PersistentOutcallContainerService passes attachments to and from
PersistentOutcallDatabase.

PersistentOutcallDatabase - I added a field called "attachment" to the
"persistentoutcalls" table.  PersistentOutcallDatabase writes
attachments to disk and saves the filename in the database, and it
reads and deletes the attachments when necessary.