Fixed a problem with aggregating requests for retrieving resource properties.
In our system, several port-types extended Enhanced-RNS and standard ByteIO port-types, but they 
do not have the relevant resource properties such as Create-Time and Size. Therefore, call 
aggregation was failing for those port-types. The problem was time-dependent as if the original 
requested resource was already in the cache, no outcall (equivalently no call aggregation) was 
taking place. This commit solves this problem by avoiding aggregation of requests when a port-type 
is an extended form of Enhanced-RNS or ByteIO port-types. Obviously, this solution comes at the 
expense of lossing a chance of performance improvement.