Initial set of changes for DLog implementation (Assuming the SVN switch works like it's supposed to. I don't know why I still have so much trouble getting SVN to cooperate!).
Port type is about half-implemented, log submit tool is done, logs can be inserted into external DB, but not yet retrieved. RPC-IDs are assigned through WorkingContext and/or CallingContetxt, depending on the code, the RPC-ID is returned with each RPC response in the SOAP Header, hierarchy is recorded in a separate table from the logs.