fix for unset bash syntax for bootstrapping, etc
not sure how this was able to escape fixing before, but i have upgraded ubuntu a few times since i last built and tested genesis II.  the issue was that bash doesn't want an equals sign ('=') at the end of an unset statement.  not sure it ever wanted one, but now it barfs if one is there.  also not sure why the '=' was there except that maybe the statement used to assign to empty ("") or something.  fixed now at least.