update to download manager to use nio for moving
the File.renameTo method was choking on some simple renames (e.g. from the job download folder to a different scratch folder in /tmp/scratch).  the nio Files.move method works in this case, and should work in other cases, even crossing partition boundaries.
the scratch tests folder is currently a non-automatable test of scratch that showed the move error.  the readme file provides more info about how to run.
besStatus test has been modified to wait for the launched asynchronous job to complete, which it was not doing before.