Two updates: Added code to copy machine to have multiple constructors depending on whether the source is a URI supporting multiple protocols, or an EPR, in which case to treat the EPR as the root of a directory structure.
I also added a new phase called "copyPersistedState" that is given an EPR of an activity living on another BES. It then
copies down the JWD contents into another a locally created JWD with the same path. It similarly copies down the accounting directory
of the job into a new accounting directory with the same path.

A few things to note.
First, for this to work ALL of the paths to the "shared" directory of JWDs and Accounting info
must be the same on every BES. We will need to go back and make sure this is so for our BESes.
Second, at this point the history records of the job are not being migrated ... we will need to set that up
and worry about whether the records indicate which BES it happened on.